requirements.txt
Django==1.10.6
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
PyJWT==1.4.2
django-crispy-forms==1.6.1
Pillow==4.1.1